unsloth/requirements.txt
# Core ML libraries
torch>=2.1.0
transformers>=4.37.0
datasets>=2.16.0
accelerate>=0.25.0
bitsandbytes>=0.42.0

# Unsloth for optimized training
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git

# TRL for fine-tuning
trl>=0.7.10
peft>=0.8.0

# Opik for experiment tracking
opik>=0.1.0

# Data processing
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0

# Environment and utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Additional utilities
matplotlib>=3.7.0
seaborn>=0.12.0

# Evaluation framework
deepeval>=0.20.0

# Jupyter (optional for development)
jupyter>=1.0.0
ipywidgets>=8.0.0 